Move components to a new sub folder
