+++ src/components/queue/view/Editor.vue
